Improved clarity by renamed some files, and rearranged parts of the code to make it more modular. Added more functionality. WIP, can't be used as it is for now

+++ get_mnist.py
@@ -8,7 +8,7 @@ import numpy as np
 
 # Although I made some changes to conform to PEP8 and add some more flexibility
 # attribution to the original code: https://github.com/hsjeong5/MNIST-for-Numpy/blob/master/mnist.py
-_filename = [
+_mnist_files = [
     ["training_images", "train-images-idx3-ubyte.gz"],
     ["test_images", "t10k-images-idx3-ubyte.gz"],
     ["training_labels", "train-labels-idx1-ubyte.gz"],
@@ -17,21 +17,28 @@ _filename = [
 
 
 def _download_mnist(data_path):
+    # TODO: add function explanation
     base_url = "http://yann.lecun.com/exdb/mnist/"
-    for name in _filename:
-        print(f"Downloading, {name[1]}. This process may take a few minutes...")
-        request.urlretrieve(base_url + name[1], data_path / name[1])
-    print("Download complete.")
+    downloaded = False
+    for name in _mnist_files:
+        file_path = data_path / name[1]
+        if not file_path.exists():
+            print(f"Downloading, {name[1]}. This process may take a few minutes...")
+            request.urlretrieve(base_url + name[1], data_path / name[1])
+            downloaded = True
+    if downloaded:
+        print("Download complete.")
 
 
 def _save_mnist(data_path):
+    # TODO: add function explanation
     mnist = {}
-    for name in _filename[:2]:
-        with gzip.open(name[1], "rb") as f:
-            mnist[name[0]] = np.frombuffer(f.read(), np.uint8, offset=16).reshape(-1, 28*28)
+    for name in _mnist_files[:2]:
+        with gzip.open(data_path / name[1], "rb") as f:
+            mnist[name[0]] = np.frombuffer(f.read(), np.uint8, offset=16).reshape(-1, 28, 28)
 
-    for name in _filename[-2:]:
-        with gzip.open(name[1], "rb") as f:
+    for name in _mnist_files[-2:]:
+        with gzip.open(data_path / name[1], "rb") as f:
             mnist[name[0]] = np.frombuffer(f.read(), np.uint8, offset=8)
 
     with open(data_path / "mnist.pkl", "wb") as f:
@@ -39,19 +46,16 @@ def _save_mnist(data_path):
     print("Save complete.")
 
 
-def load_mnist(data_path=Path("data")):
-    if not data_path.exists():
-        data_path.mkdir()
+def load_mnist(data_path):
+    # TODO: add function explanation
     if not data_path.is_dir():
         raise ValueError("Data path is not a directory")
 
-    mnist_file = data_path / "mnist.pkl"
-    if not mnist_file.exists():
-        print("No existing mnist file found, download will start now.")
+    mnist_pkl = data_path / "mnist.pkl"
+    if not mnist_pkl.exists():
         _download_mnist(data_path)
         _save_mnist(data_path)
-    with open(mnist_file, "rb") as f:
-        print("Existing mnist file found, proceeding to load.")
+    with open(mnist_pkl, "rb") as f:
         mnist = pickle.load(f)
 
     return mnist["training_images"], mnist["training_labels"], mnist["test_images"], mnist["test_labels"]
